requirements.txt
gradio
torch
torchvision
huggingface_hub
Pillow
